feat: ship ModernBERT embeddings and Qwen3.5 2B catalog entries
Bump the download catalog to schema version 2 and publish two accepted
model entries: the 8-bit affine ModernBERT embedding encoder and the
4-bit Qwen3.5 2B chat model, pinned to immutable revisions. The catalog
family classification learns ModernBERT with a dedicated embedding
capability badge, the Observatory renders the new family and badge, and
the bundled-catalog acceptance tests pin both revisions and purpose
capabilities. The Observatory console test suite is split by concern so
each file stays within the source-size contract.
