+++ apps/supervisor/tests/hermetic/download_manifest_preflight.rs
@@ -325,7 +325,7 @@ fn json_response(body: serde_json::Value) -> HubHttpResponse {
 
 fn catalog_json() -> String {
     format!(
-        "{{\"schema_version\":1,\"entries\":[{{\"huggingface_id\":\"{REPOSITORY_ID}\",\"revision\":\"{REVISION}\",\"display_name\":\"Example model\",\"family\":\"qwen3_5\",\"approximate_size_bytes\":100,\"public\":true}}]}}"
+        "{{\"schema_version\":2,\"entries\":[{{\"huggingface_id\":\"{REPOSITORY_ID}\",\"revision\":\"{REVISION}\",\"display_name\":\"Example model\",\"family\":\"qwen3_5\",\"approximate_size_bytes\":100,\"public\":true}}]}}"
     )
 }
