+++ repo-discovery-guide-for-agents.md
@@ -4,7 +4,7 @@ Agent orientation map for non-obvious repository facts; this is not user-facing 
 
 ## Maintenance mandate
 
-Read this guide before substantive work and spot-check 2-3 key facts. Update it when paths, entry points, conventions, or expensive gotchas change. Treat it as suspect when Last verified is more than 90 days old. Last verified: 2026-09-05 (native `POST /v1/embeddings` on ModernBERT 8-bit; MLX pin v0.32.2).
+Read this guide before substantive work and spot-check 2-3 key facts. Update it when paths, entry points, conventions, or expensive gotchas change. Treat it as suspect when Last verified is more than 90 days old. Last verified: 2026-09-06 (catalog schema 2 adds ModernBERT embeddings entries; native `POST /v1/embeddings` on ModernBERT 8-bit; MLX pin v0.32.2).
 
 ## Project overview
 
@@ -18,7 +18,7 @@ Astronomical is an Apple Silicon local model server. The active architecture is 
 - The local API supports POST /v1/chat/completions for OpenAI chat chunks, POST /v1/responses for Copilot-compatible response items, POST /v1/images/generations for one lossless base64 PNG, and POST /v1/embeddings for one unit-norm vector per input. Chat and Responses share the autoregressive worker command; image generation and embeddings each have a separate typed command. Embeddings are synchronous single-shot: no step loop and no mid-request cancellation.
 - Models are loaded lazily: menu-app startup launches an idle worker with no resident model. The supervisor discovers configured models for /v1/models, canonicalizes supported provider-prefixed IDs, then sends SwapModel only when the first (or a different) REST generation request targets a discovered model. It waits for ModelSwapped before forwarding generation. Process-scoped telemetry and configuration events can precede that acknowledgement and must pass through the shared typed event handler; generation-scoped events remain protocol violations. A rejected model load leaves the worker idle or preserves the previous model so a later valid request can proceed.
 - Every instance automatically scans `<instance-state>/models` before user-configured recursive roots. An absent automatic root is an empty Library and is not created during resolution. Effective roots preserve automatic-then-authored order and deduplicate exact lexically equal paths, while `ResolvedRuntimeConfig.configured_model_directories` retains only user-authored roots. A Library model wins over the same public identity in an authored root. Ambiguous identities among authored roots are excluded while unrelated models remain available. An unreadable authored root is skipped with a path-free `unavailable_model_directory` diagnostic so remaining models still start; the menu retries daemon startup instead of sitting on a dead port. Bounded status diagnostics identify the model and one-based configuration entries without exposing paths. Existing organization/model trees are discoverable; hidden `.incomplete` staging remains undiscoverable.
-- `registry/download_catalog.json` is an immutable release-authored catalog embedded into astronomicald and validated before listener binding. Catalog metadata is bounded to 1,000,000 bytes, 1,024 entries, and 256 bytes per display name. Schema version 1 accepts executable Qwen3.5, Laguna, and FLUX.2 Klein entries with safe public Hugging Face identities, immutable lowercase 40-character hexadecimal revisions, JavaScript-safe byte sizes, optional bounded executable-path selections, and no unknown fields or case-insensitive collisions. `GET /v1/library/catalog` serves authored order and live readiness without worker dependency. Observatory `/library` retries interrupted catalog loads, renders decimal SI sizes, and controls one daemon-owned download.
+- `registry/download_catalog.json` is an immutable release-authored catalog embedded into astronomicald and validated before listener binding. Catalog metadata is bounded to 1,000,000 bytes, 1,024 entries, and 256 bytes per display name. Schema version 2 accepts executable Qwen3.5, Laguna, FLUX.2 Klein, and ModernBERT embedding entries with safe public Hugging Face identities, immutable lowercase 40-character hexadecimal revisions, JavaScript-safe byte sizes, optional bounded executable-path selections, capability badges including `supports_embeddings`, and no unknown fields or case-insensitive collisions. `GET /v1/library/catalog` serves authored order and live readiness without worker dependency. Observatory `/library` retries interrupted catalog loads, renders decimal SI sizes, and controls one daemon-owned download.
 - One durable Library transfer uses `<instance-state>/models/.download-job.json`; independent store instances serialize mutations through `.download-job.lock`, and partial files remain under `.incomplete/organization/model/`. Job metadata is bounded, strict, atomically replaced, and reconciled from synchronized regular-file lengths after restart. Manifest preflight derives both disk checks from this models volume and persists only release-selected executable files with exact provider digests. Payload transfer streams one immutable file range at a time; quarter-second user-interface progress stays in memory so synchronized job rewrites do not throttle the network path. File completion, pause, failure, verification, and publication remain durable boundaries. Provider SHA-256 or Git-blob SHA-1 evidence is verified before staging is atomically renamed. Publication writes strict provider identity and immutable revision provenance so ordinary discovery preserves readiness after daemon restart. A pre-existing destination is adopted only when every selected file matches provider digest evidence; mismatches remain untouched. A checksum failure truncates only the invalid file so Resume re-fetches it while retaining verified progress. Interrupted preflight, manifest, transfer, and verification states recover as paused. Cancel removes only validated hidden staging and the job document. After rename, readiness remains publishing until discovery refresh succeeds, and restart retries that refresh.
 - `/v1/models` publishes maximum input and output token counts as independent capabilities under one shared context window. The maximum input reserves one generation position; request admission enforces the selected prompt plus output budget. Never reject metadata merely because the two independent maxima cannot be used simultaneously.
 - Chat, image, and embeddings requests share one serial bounded FIFO queue and one active-request permit. Up to GENERATION_QUEUE_DEPTH requests can wait; excess requests receive HTTP 429 with server_capacity. Queue permits are released immediately when a queued request is cancelled or its HTTP client disconnects.
