+++ scripts/verify-before-commit.sh
@@ -118,7 +118,8 @@ main() {
         --test --test-reporter=spec .github/scripts/pull-request-issue-compliance.test.js
     run_step test-observatory-contracts "$TEST_TIMEOUT_SECONDS" node --test --test-reporter=spec \
         apps/supervisor/console/console.test.js \
-        apps/supervisor/console/library.test.js
+        apps/supervisor/console/library.test.js \
+        apps/supervisor/console/library-fetch.test.js
     run_step compile-rust "$COMPILE_TIMEOUT_SECONDS" cargo verify-commit-rust \
         --timings --no-run --jobs "$logical_cpu_count"
     run_step run-rust "$TEST_TIMEOUT_SECONDS" cargo verify-commit-rust \
